tache du 23/10/2021

+++ src/Controller/superAdmin/CategoryController.php
@@ -6,6 +6,7 @@ use App\Entity\Categories;
 use App\Entity\Category;
 use App\Form\CategoriesType;
 use App\Form\CategoryType;
+use App\Repository\ArticleRepository;
 use App\Repository\CategoryRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -23,12 +24,21 @@ class CategoryController extends AbstractController
     /**
      * @Route("/new", name="add_category")
      */
-    public function add( Request $request , EntityManagerInterface $em): Response
+    public function add( Request $request , EntityManagerInterface $em, CategoryRepository $categoryRepository): Response
     {
         $categories = new Category();
         $form = $this->createForm(CategoryType::class,$categories);
         $form->handleRequest($request);
         if($form->isSubmitted() && $form->isValid()){
+            $categoryExiste = $categoryRepository->findBy(array('title' =>  $form->get('title')->getData()));
+            if ($categoryExiste) {
+                $this->addFlash('error','La catégorie '.$form->get('title')->getData().' existe');
+
+                return $this->render('superAdmin/category/new.html.twig',[
+                    'form' => $form->createView(),
+                    'categorie' => ''
+                ]);
+            }
             $em->persist($categories);
             $em->flush();
             $this->addFlash('success','Ajout effectué avec succés');
@@ -59,9 +69,16 @@ class CategoryController extends AbstractController
     /**
      * @Route("/{id<\d+>}", name="delete_category")
      */
-    public function delete(Category $categories, EntityManagerInterface $em, CategoryRepository $categoryRepository): Response
+    public function delete(Category $categories, EntityManagerInterface $em, CategoryRepository $categoryRepository, ArticleRepository $articleRepository): Response
     {
         if ($categories) {
+            $categorieLieByArticle = $articleRepository->findBy(array('categorie' => $categories));
+
+            if ($categorieLieByArticle) {
+                $this->addFlash('error','Cet catégorie lié avec un ou des (articles), tu ne peut pas supprimers');
+                return $this->redirectToRoute('index_category');
+            }
+
             $em->remove($categories);
             $em->flush();
             $this->addFlash('success','Supprimer effectué avec succés');
@@ -75,14 +92,16 @@ class CategoryController extends AbstractController
     /**
      * @Route("/edit/{id<\d+>}", name="edit_category")
      */
-    public function edit(Category $categorie, Request $request, EntityManagerInterface $em): Response
+    public function edit(Category $categorie, Request $request, EntityManagerInterface $em, ArticleRepository $articleRepository): Response
     {
 
 
 
         $form = $this->createForm(CategoryType::class,$categorie);
         $form->handleRequest($request);
+
         if($form->isSubmitted() && $form->isValid()){
+
             $em->persist($categorie);
             $em->flush();
             $this->addFlash('success','Modifier effectué avec succés');
